UUR/requirements.txt
python-dotenv==1.1.0
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.4
elevenlabs==2.1.0
groq==0.28.0